+++ ton_client/src/net/errors.rs
@@ -2,19 +2,18 @@ use crate::error::ClientError;
 use serde_json::Value;
 use std::fmt::Display;
 
-const NET: isize = ClientError::NET; // 600
-
+#[derive(ApiType)]
 pub enum ErrorCode {
-    QueryFailed = NET + 1,
-    SubscribeFailed = NET + 2,
-    WaitForFailed = NET + 3,
-    GetSubscriptionResultFailed = NET + 4,
-    InvalidServerResponse = NET + 5,
-    ClockOutOfSync = NET + 6,
-    WaitForTimeout = NET + 7,
-    GraphqlError = NET + 8,
-    NetworkModuleSuspended = NET + 9,
-    WebsocketDisconnected = NET + 10,
+    QueryFailed = 601,
+    SubscribeFailed = 602,
+    WaitForFailed = 603,
+    GetSubscriptionResultFailed = 604,
+    InvalidServerResponse = 605,
+    ClockOutOfSync = 606,
+    WaitForTimeout = 607,
+    GraphqlError = 608,
+    NetworkModuleSuspended = 609,
+    WebsocketDisconnected = 610,
 }
 
 pub struct Error;