+++ CHANGELOG.md
@@ -14,6 +14,7 @@ All notable changes to this project will be documented in this file.
 - `tonclient-core-version` http header.
 - `net.find_last_shard_block` function returning account shard last block ID.
 - `boc.get_code_from_tvc` function extracting contract code from TVC image.
+- `ErrorCode` type in each module spec in `api.json`.
 - **Debot Module:**
   - Add new variant `ParamsOfAppDebotBrowser::SwitchCompleted` to notify browser when all context actions are shown.
   - Added new 3 engine routines for crypto operations and 1 routine for querying account state (balance, state type, code, data) that can be used in debots.